+++ src/aks-preview/azext_aks_preview/custom.py
@@ -596,6 +596,7 @@ def aks_create(
     guardrails_excluded_ns=None,
     # azure service mesh
     enable_azure_service_mesh=None,
+    revision=None,
     # azure monitor profile
     enable_azuremonitormetrics=False,
     enable_azure_monitor_metrics=False,